requirements.txt
Flask-Mysql
flask-table
flask_cors
flask_jwt_extended
flask_httpauth
